fix: make public SkillHub build self-contained
