+++ docs-site/package.json
@@ -8,8 +8,8 @@
   },
   "scripts": {
     "dev": "vitepress dev docs",
-    "skillhub:build": "node ../scripts/build-public-skillhub.mjs",
-    "skillhub:check": "node ../scripts/build-public-skillhub.mjs --check",
+    "skillhub:build": "node scripts/build-public-skillhub.mjs",
+    "skillhub:check": "node scripts/build-public-skillhub.mjs --check",
     "prebuild": "npm run skillhub:build",
     "build": "vitepress build docs",
     "preview": "vitepress preview docs"
